Adapt link indexing to order in gamma_in
The links in x direction are counted first in x, then in y direction.
The links in y direction are counted first in y, then in x direction.
This leads to an easier order in the modes. As long as the lattice is
asked to do the indexing, everything can be expressed in coordinates.